\name{print.clust}
\alias{print.clust}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Print method for object clust }
\description{
  Print method for object \sQuote{clust}.
}
\usage{
print.clust(x, ...)
}
\arguments{
  \item{x}{ Object from function \sQuote{clust} }
 \item{\dots}{  }
}
\author{ Matthias Templ }
\seealso{ \code{\link{clust} }}
\examples{
  GUIspatClust()
}
\keyword{ print }
